Fixed an issue with the recently added function
The issue was that it printed done and undone with the same color (red).
This behavior is obviously not what it should be doing, so I added
a check for item->done to add the color based on that.
Note that this check is performed twice because it is also what
determines wether we print a cross (no) or a checkmark (yes); but
I don't know if there's a way to do it without checking that twice. Well
if there is a way, then it's not using printf surely, and it's not
really worth the trouble.